Completely overhauled the createCV script & renamed Create-CustomViews.ps1
The script has been effectively rewritten in a more "natural" .NET/PoSh
style; the previous version was written using very "C-style" semantics.

Note that the support for creating Custom Views filtered to a given time
has been temporarily removed until the associated code is also rewritten.

The end result is a script that is shorter, more concise & easier to read.