Initial version of Create-ScheduledTasks (incomplete!!)
This is not yet suitable for use but will allow for creating corresponding
event triggered scheduled tasks based on WEC subscriptions. A command can
thus be triggered on arrival of a matching event to, for example, send an
email notifying an Ops team or perform some other desired action. This can
substantially increase the utility of log aggregation via automating the
responses to given classes of events.